fix(repl): disallow shorthand by default
Options were being parsed with shorthand enabled by default; that is, as being able to be referenced with shorthand by the first letter of their name. The CLI may face issues whenever two options would start with the same letter, were this to remain the case; options with shorthand capability should have it enabled explicitly.

This commit also ensures that 0-argument options are given the single argument "SET", so that such options are treated properly as flags.