feat(repl): add command preregistry
Preregistration of commands allows for the program to be AWARE of all possible commands, while still barring users from using some of them as before. EG: the 'account' command is dependent on the user first viewing a budget whose accounts they would like to see, and the 'budget' command is dependent on them logging in first. With pregistration, we can avoid insisting to the user that these accounts are unknown, instead letting them know the course of action required to use them via the command handlers.